feat: 라우터 설정 및 input 커스텀훅 수정
