+++ apps/client/src/pages/community/community-page.css.ts
@@ -1,1 +1,10 @@
-// import { style } from '@vanilla-extract/css';
+import { style } from '@vanilla-extract/css';
+
+export const emptyPlaceholder = style({
+  display: 'flex',
+  flex: 1,
+  width: '100%',
+  height: '100%',
+  alignItems: 'center',
+  justifyContent: 'center',
+});
