+++ apps/client/src/pages/community/community-write/community-write.css.ts
@@ -1,10 +1,27 @@
 import { style } from '@vanilla-extract/css';
 
-export const emptyPlaceholder = style({
+export const container = style({
   display: 'flex',
-  flex: 1,
-  width: '100%',
+  flexDirection: 'column',
   height: '100%',
-  alignItems: 'center',
-  justifyContent: 'center',
+  gap: '3.6rem',
+});
+
+export const postContainer = style({
+  display: 'flex',
+  flexDirection: 'column',
+  padding: '0 1.6rem',
+  gap: '3.6rem',
+});
+
+export const postHeader = style({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '1.2rem',
+});
+
+export const postContent = style({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '1.2rem',
 });
