+++ apps/client/src/pages/community/community-write/community-write.tsx
@@ -1,31 +1,80 @@
-import { Alert, Floating, Navigation } from '@bds/ui';
+import { useEffect, useState } from 'react';
+import { Input, Navigation, TextButton, Title } from '@bds/ui';
 import { Icon } from '@bds/ui/icons';
+import { useNavigate } from 'react-router';
 
-import EmptyPlaceholder from '@widgets/community/components/empty-placeholder/empty-placeholder';
-import { ALERT_CONTENT_BODY } from '@widgets/community/constant/alert_content';
-import { EMPTY_POST } from '@widgets/community/constant/empty-content';
+import CommunityLine from '@widgets/community/components/community-line/community-line';
+
+import { LIMIT_SHORT_TEXT } from '@shared/constants/text_limits';
+import { useInputState } from '@shared/hooks/use-input-state';
+import { useLimitedInput } from '@shared/hooks/use-limited-input';
+import { useTextAreaState } from '@shared/hooks/use-textarea-state';
 
 import * as styles from './community-write.css';
 
+const COMMUNITY_CONTENT = {
+  TITLE: {
+    HEADER: '제목',
+    BODY: '내용',
+  },
+  BUTTON: '게시물 올리기',
+};
+
 const CommunityWrite = () => {
+  const [title, onTitleChange] = useInputState('', (v) => v.trim());
+  const [content, onContentChange] = useTextAreaState();
+  const [isDisabled, setIsDisabled] = useState(true);
+  const navigation = useNavigate();
+  const { isErrorState } = useLimitedInput(LIMIT_SHORT_TEXT, title.length);
+
+  useEffect(() => {
+    const isTitleValid = title.trim().length > 0;
+    const isContentValid = content.trim().length > 0;
+
+    setIsDisabled(!(isTitleValid && isContentValid));
+  }, [title, content]);
+
+  const handleGoBack = () => {
+    navigation(-1);
+  };
+
   return (
-    <div>
-      <Navigation rightIcon={<Icon name="home" />} title="커뮤니티" />
-      <Alert
-        iconName="info"
-        iconSize="2.4rem"
-        alertHeader={ALERT_CONTENT_BODY.HEADER}
-        alertContents={ALERT_CONTENT_BODY.BODY}
-        type="info"
+    <div className={styles.container}>
+      <Navigation
+        title="글쓰기"
+        leftIcon={
+          <Icon
+            name="caret_left_lg"
+            width="2.4rem"
+            height="2.4rem"
+            onClick={handleGoBack}
+          />
+        }
+        rightIcon={
+          <TextButton
+            color="primary"
+            disabled={isDisabled}
+            // @POST OnClick
+          >
+            올리기
+          </TextButton>
+        }
       />
-      <div className={styles.emptyPlaceholder}>
-        <EmptyPlaceholder content={EMPTY_POST} />
+      <div className={styles.postContainer}>
+        <div className={styles.postHeader}>
+          <Title fontStyle="eb_md">{COMMUNITY_CONTENT.TITLE.HEADER}</Title>
+          <Input
+            value={title}
+            onChange={onTitleChange}
+            bgColor="gray"
+            errorState={isErrorState}
+          />
+        </div>
+        <div className={styles.postContent}>
+          <Title fontStyle="eb_md">{COMMUNITY_CONTENT.TITLE.BODY}</Title>
+          <CommunityLine value={content} onChange={onContentChange} />
+        </div>
       </div>
-      <Floating
-        icon={<Icon name="edit" width={'100%'} height={'100%'} />}
-        state="default"
-        //onClick to @커뮤니티_글_작성 화면
-      />
     </div>
   );
 };
